fixed problems

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -9,7 +9,7 @@
  * Return: 1 on success, -1 on failure
  */
 
-int clear_bit(unsigned int *n, unsigned int index)
+int clear_bit(unsigned long *n, unsigned int index)
 {
 	if (index > sizeof(int) * 8)
 		return (-1);
